courses/UCLC1008/AIagentDemo/MockPractice/answerSheet/requirements.txt
pymupdf>=1.23.0
Pillow>=9.0.0
numpy>=1.20.0
